+++ miniapp/src/main/java/com/rakuten/tech/mobile/miniapp/js/MessageToContact.kt
@@ -8,10 +8,9 @@ data class MessageToContact(
     val image: String,
     val text: String,
     val caption: String,
-    val action: String,
-    val miniAppTitle: String
+    val action: String
 ) {
     /** Returns true when all the properties are empty, otherwise returns false. */
     val isEmpty: Boolean =
-        image.isEmpty() && text.isEmpty() && caption.isEmpty() && action.isEmpty() && miniAppTitle.isEmpty()
+        image.isEmpty() && text.isEmpty() && caption.isEmpty() && action.isEmpty()
 }