+++ testapp/src/main/java/com/rakuten/tech/mobile/testapp/ui/chat/ChatWindow.kt
@@ -2,14 +2,14 @@ package com.rakuten.tech.mobile.testapp.ui.chat
 
 import android.app.Activity
 import android.view.LayoutInflater
+import android.widget.Toast
 import androidx.appcompat.app.AlertDialog
 import androidx.recyclerview.widget.DividerItemDecoration
 import androidx.recyclerview.widget.LinearLayoutManager
 import com.rakuten.tech.mobile.miniapp.js.MessageToContact
 import com.rakuten.tech.mobile.miniapp.js.userinfo.Contact
 import com.rakuten.tech.mobile.miniapp.testapp.R
-import com.rakuten.tech.mobile.miniapp.testapp.databinding.DialogContactMessageContentBinding
-import com.rakuten.tech.mobile.miniapp.testapp.databinding.WindowContactSelectionBinding
+import com.rakuten.tech.mobile.miniapp.testapp.databinding.WindowChatBinding
 import com.rakuten.tech.mobile.testapp.helper.load
 import com.rakuten.tech.mobile.testapp.helper.showAlertDialog
 import com.rakuten.tech.mobile.testapp.ui.settings.AppSettings
@@ -17,22 +17,23 @@ import kotlinx.coroutines.Dispatchers
 import kotlinx.coroutines.GlobalScope
 import kotlinx.coroutines.launch
 
-class ContactSelectionWindow(private val activity: Activity) :
+class ChatWindow(private val activity: Activity) :
     ContactSelectionAdapter.ContactSelectionListener {
 
     lateinit var contactSelectionAlertDialog: AlertDialog
-    lateinit var messageSentAlertDialog: AlertDialog
     private val layoutInflater = LayoutInflater.from(activity)
     private lateinit var contactSelectionAdapter: ContactSelectionAdapter
 
     private val hasContact =
         AppSettings.instance.isContactsSaved && !AppSettings.instance.contacts.isNullOrEmpty()
 
+    private lateinit var miniAppTitle: String
     private lateinit var message: MessageToContact
     private lateinit var onSuccessSingleContact: (contactId: String?) -> Unit
     private lateinit var onErrorSingleContact: (message: String) -> Unit
 
     fun openSingleContactSelection(
+        miniAppTitle: String,
         message: MessageToContact,
         onSuccess: (contactId: String?) -> Unit,
         onError: (message: String) -> Unit
@@ -41,16 +42,16 @@ class ContactSelectionWindow(private val activity: Activity) :
             showAlertDialog(
                 activity,
                 "Warning",
-                "There is no contact found saved in HostApp."
+                "There is no contact found saved in HostApp!"
             )
             return
         }
 
+        this.miniAppTitle = miniAppTitle
         this.message = message
         this.onSuccessSingleContact = onSuccess
         this.onErrorSingleContact = onError
 
-        // initialize contact selection view
         initDefaultWindow()
         prepareDataForAdapter()
 
@@ -59,19 +60,32 @@ class ContactSelectionWindow(private val activity: Activity) :
     }
 
     private fun initDefaultWindow() {
-        val contactView = WindowContactSelectionBinding.inflate(layoutInflater, null, false)
-        contactView.listContactSelection.layoutManager = LinearLayoutManager(activity)
-        contactView.listContactSelection.addItemDecoration(
+        val rootView = WindowChatBinding.inflate(layoutInflater, null, false)
+
+        // set message content
+        GlobalScope.launch(Dispatchers.Main) {
+            message.apply {
+                rootView.messageImage.load(activity, image)
+                rootView.messageText.text = text
+                rootView.miniAppTitle.text = miniAppTitle
+                rootView.messageCaption.text = caption // todo
+                rootView.messageAction.text = action // todo
+            }
+        }
+
+        // set list of contacts to select
+        rootView.listContactSelection.layoutManager = LinearLayoutManager(activity)
+        rootView.listContactSelection.addItemDecoration(
             DividerItemDecoration(activity, DividerItemDecoration.VERTICAL)
         )
 
         contactSelectionAdapter = ContactSelectionAdapter(this)
-        contactView.listContactSelection.adapter = contactSelectionAdapter
+        rootView.listContactSelection.adapter = contactSelectionAdapter
 
         contactSelectionAlertDialog =
             AlertDialog.Builder(activity, R.style.AppTheme_DefaultWindow).create()
-        contactSelectionAlertDialog.setView(contactView.root)
-        contactView.contactCloseWindow.setOnClickListener {
+        contactSelectionAlertDialog.setView(rootView.root)
+        rootView.chatActionCancel.setOnClickListener {
             contactSelectionAlertDialog.dismiss()
         }
     }
@@ -89,34 +103,12 @@ class ContactSelectionWindow(private val activity: Activity) :
             else -> {
                 onSuccessSingleContact(contact.id)
                 // Note: Doesn't need to actually send a message because we don't have an interface for this in the demo app.
-                showMessageDialog(contact.id)
+                Toast.makeText(
+                    activity, "The message has been sent to contact id: ${contact.id}", Toast.LENGTH_LONG
+                ).show()
             }
         }
 
         contactSelectionAlertDialog.dismiss()
-    }
-
-    private fun showMessageDialog(contactId: String) {
-        // set message attributes to views
-        val messageView = DialogContactMessageContentBinding.inflate(layoutInflater, null, false)
-
-        GlobalScope.launch(Dispatchers.Main) {
-            message.apply {
-                messageView.messageImage.load(activity, image)
-                messageView.messageText.text = text
-                messageView.messageTitle.text = miniAppTitle
-                messageView.messageCaption.text = caption
-                messageView.messageAction.text = action
-            }
-            messageView.messageGeneric.text = "The message has been sent to contact id: ${contactId}"
-
-            // set dialog
-            messageSentAlertDialog = AlertDialog.Builder(activity).create()
-            messageSentAlertDialog.setView(messageView.root)
-            messageView.dialogDismiss.setOnClickListener {
-                messageSentAlertDialog.dismiss()
-            }
-            messageSentAlertDialog.show()
-        }
     }
 }