+++ testapp/src/main/java/com/rakuten/tech/mobile/testapp/ui/display/MiniAppDisplayActivity.kt
@@ -29,7 +29,7 @@ import com.rakuten.tech.mobile.miniapp.testapp.R
 import com.rakuten.tech.mobile.miniapp.testapp.databinding.MiniAppDisplayActivityBinding
 import com.rakuten.tech.mobile.testapp.helper.AppPermission
 import com.rakuten.tech.mobile.testapp.ui.base.BaseActivity
-import com.rakuten.tech.mobile.testapp.ui.chat.ContactSelectionWindow
+import com.rakuten.tech.mobile.testapp.ui.chat.ChatWindow
 import com.rakuten.tech.mobile.testapp.ui.settings.AppSettings
 import java.util.*
 
@@ -40,6 +40,7 @@ class MiniAppDisplayActivity : BaseActivity() {
     private var miniappPermissionCallback: (isGranted: Boolean) -> Unit = {}
     private lateinit var sampleWebViewExternalResultHandler: ExternalResultHandler
     private lateinit var binding: MiniAppDisplayActivityBinding
+    private var miniAppTitle = ""
 
     private val externalWebViewReqCode = 100
     private val fileChoosingReqCode = 10101
@@ -91,6 +92,7 @@ class MiniAppDisplayActivity : BaseActivity() {
         val appInfo = intent.getParcelableExtra<MiniAppInfo>(miniAppTag)
         val appId = intent.getStringExtra(appIdTag) ?: appInfo?.id
         val appUrl = intent.getStringExtra(appUrlTag)
+        miniAppTitle = appInfo?.displayName.toString()
 
         binding = DataBindingUtil.setContentView(this, R.layout.mini_app_display_activity)
 
@@ -208,6 +210,7 @@ class MiniAppDisplayActivity : BaseActivity() {
         miniAppMessageBridge.setUserInfoBridgeDispatcher(userInfoBridgeDispatcher)
 
         // setup ChatBridgeDispatcher
+        val chatWindow = ChatWindow(this@MiniAppDisplayActivity)
         val chatBridgeDispatcher = object : ChatBridgeDispatcher {
 
             override fun sendMessageToContact(
@@ -215,8 +218,15 @@ class MiniAppDisplayActivity : BaseActivity() {
                 onSuccess: (contactId: String?) -> Unit,
                 onError: (message: String) -> Unit
             ) {
-                val contactSelectionWindow = ContactSelectionWindow(this@MiniAppDisplayActivity)
-                contactSelectionWindow.openSingleContactSelection(message, onSuccess, onError)
+                chatWindow.openSingleContactSelection(miniAppTitle, message, onSuccess, onError)
+            }
+
+            override fun sendMessageToMultipleContacts(
+                message: MessageToContact,
+                onSuccess: (contactIds: List<String>) -> Unit,
+                onError: (message: String) -> Unit
+            ) {
+                //chatWindow.openSingleContactSelection(miniAppTitle, message, onSuccess, onError)
             }
         }
         miniAppMessageBridge.setChatBridgeDispatcher(chatBridgeDispatcher)